core/src/test/java/org/opensearch/graphql/translation/EnumTypeTranslationTest.java
package org.opensearch.graphql.translation;

import graphql.schema.GraphQLSchema;
import graphql.schema.idl.EchoingWiringFactory;
import org.junit.Assert;
import org.junit.jupiter.api.AfterAll;
import org.junit.jupiter.api.BeforeAll;
import org.junit.jupiter.api.Test;
import org.opensearch.graphql.GraphQLEngineFactory;
import org.opensearch.schema.ontology.Accessor;
import org.opensearch.schema.ontology.Value;

import java.io.FileInputStream;
import java.io.FileNotFoundException;
import java.io.InputStream;
import java.util.Arrays;
import java.util.List;
import java.util.stream.Collectors;

class EnumTypeTranslationTest {
    static GraphQLSchema schema;

    @AfterAll
    public static void tearDown() throws Exception {
        GraphQLEngineFactory.reset();
    }

    @BeforeAll
    static void setup() throws FileNotFoundException {
        InputStream utilsSchemaInput = new FileInputStream("../schema/utils.graphql");
        InputStream stream = Thread.currentThread().getContextClassLoader().getResourceAsStream("graphql/sample/simpleGQLForeignBooks.graphql");
        schema = GraphQLEngineFactory.generateSchema(new EchoingWiringFactory(), Arrays.asList(utilsSchemaInput,stream));
    }

    @Test
    void translate() {
        TranslationStrategy.TranslationContext context = new TranslationStrategy.TranslationContext("test");
        ObjectTypeTranslation objectTypeTranslation = new ObjectTypeTranslation();
        objectTypeTranslation.translate(schema, context);

        EnumTypeTranslation transformer = new EnumTypeTranslation();
        transformer.translate(schema, context);
        Accessor accessor = new Accessor(context.getBuilder().build());
        //verify enum created
        Assert.assertTrue(accessor.enumeratedType("genre").isPresent());
        List<String> collect = accessor.enumeratedType("genre").get().getValues().stream().map(Value::getName).collect(Collectors.toList());
        Assert.assertEquals(List.of("AdventureStories", "Classics", "Crime", "FairyTales", "Fantasy", "HistoricalFiction", "Horror", "Humour"), collect);
    }
}